+++ src/errors.rs
@@ -11,3 +11,9 @@ impl serde::de::Error for RealmError {
         Self::Anyhow(anyhow::anyhow!(msg.to_string()))
     }
 }
+
+impl serde::ser::Error for RealmError {
+    fn custom<T: std::fmt::Display>(msg: T) -> Self {
+        Self::Anyhow(anyhow::anyhow!(msg.to_string()))
+    }
+}