+++ src/value/ser.rs
@@ -1,12 +1,456 @@
-// use serde::{Serialize, Serializer};
+use serde::ser::Serialize;
 
-// use super::Value;
+use super::Value;
+use crate::{map::Map, RealmError};
 
-// impl Serialize for Value {
-//     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-//     where
-//         S: Serializer,
-//     {
-//         todo!()
-//     }
-// }
+pub struct ValueSerializer;
+
+impl<'a> serde::ser::Serializer for &'a mut ValueSerializer {
+    type Ok = Value;
+    type Error = RealmError;
+
+    type SerializeSeq = SerializeSeq;
+    type SerializeTuple = SerializeSeq;
+    type SerializeTupleStruct = SerializeSeq;
+    type SerializeTupleVariant = SerializeSeq;
+    type SerializeMap = SerializeMap;
+    type SerializeStruct = SerializeMap;
+    type SerializeStructVariant = SerializeMap;
+
+    fn serialize_bool(self, v: bool) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Boolean(v))
+    }
+
+    fn serialize_i8(self, v: i8) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Integer(v.into()))
+    }
+
+    fn serialize_i16(self, v: i16) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Integer(v.into()))
+    }
+
+    fn serialize_i32(self, v: i32) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Integer(v.into()))
+    }
+
+    fn serialize_i64(self, v: i64) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Integer(v))
+    }
+
+    fn serialize_u8(self, v: u8) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Integer(v.into()))
+    }
+
+    fn serialize_u16(self, v: u16) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Integer(v.into()))
+    }
+
+    fn serialize_u32(self, v: u32) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Integer(v.into()))
+    }
+
+    fn serialize_u64(self, v: u64) -> Result<Self::Ok, Self::Error> {
+        if i64::try_from(v).is_ok() {
+            Ok(Value::Integer(v as i64))
+        } else {
+            Err(serde::ser::Error::custom("u64 value convert failed"))
+        }
+    }
+
+    fn serialize_f32(self, v: f32) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Float(v.into()))
+    }
+
+    fn serialize_f64(self, v: f64) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Float(v))
+    }
+
+    fn serialize_char(self, v: char) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::String(v.to_string()))
+    }
+
+    fn serialize_str(self, v: &str) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::String(v.to_string()))
+    }
+
+    fn serialize_bytes(self, v: &[u8]) -> Result<Self::Ok, Self::Error> {
+        let vec = v.iter().map(|&b| Value::Integer(b.into())).collect();
+        Ok(Value::Array(vec))
+    }
+
+    fn serialize_none(self) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Null)
+    }
+
+    fn serialize_some<T>(self, value: &T) -> Result<Self::Ok, Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        value.serialize(self)
+    }
+
+    fn serialize_unit(self) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Null)
+    }
+
+    fn serialize_unit_struct(
+        self,
+        _name: &'static str,
+    ) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Null)
+    }
+
+    fn serialize_unit_variant(
+        self,
+        _name: &'static str,
+        _variant_index: u32,
+        variant: &'static str,
+    ) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::String(variant.to_string()))
+    }
+
+    fn serialize_newtype_struct<T>(
+        self,
+        _name: &'static str,
+        value: &T,
+    ) -> Result<Self::Ok, Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        value.serialize(self)
+    }
+
+    fn serialize_newtype_variant<T>(
+        self,
+        _name: &'static str,
+        _variant_index: u32,
+        variant: &'static str,
+        value: &T,
+    ) -> Result<Self::Ok, Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        let mut map = Map::new();
+        map.insert(variant.to_string(), value.serialize(self)?);
+        Ok(Value::Table(map))
+    }
+
+    fn serialize_seq(
+        self,
+        len: Option<usize>,
+    ) -> Result<Self::SerializeSeq, Self::Error> {
+        Ok(SerializeSeq {
+            vec: Vec::with_capacity(len.unwrap_or(0)),
+        })
+    }
+
+    fn serialize_tuple(
+        self,
+        len: usize,
+    ) -> Result<Self::SerializeTuple, Self::Error> {
+        self.serialize_seq(Some(len))
+    }
+
+    fn serialize_tuple_struct(
+        self,
+        _name: &'static str,
+        len: usize,
+    ) -> Result<Self::SerializeTupleStruct, Self::Error> {
+        self.serialize_seq(Some(len))
+    }
+
+    fn serialize_tuple_variant(
+        self,
+        _name: &'static str,
+        _variant_index: u32,
+        variant: &'static str,
+        len: usize,
+    ) -> Result<Self::SerializeTupleVariant, Self::Error> {
+        let mut map = Map::new();
+        map.insert(variant.to_string(), Value::Array(Vec::with_capacity(len)));
+        Ok(SerializeSeq {
+            vec: Vec::with_capacity(len),
+        })
+    }
+
+    fn serialize_map(
+        self,
+        _len: Option<usize>,
+    ) -> Result<Self::SerializeMap, Self::Error> {
+        Ok(SerializeMap {
+            map: Map::new(),
+            next_key: None,
+        })
+    }
+
+    fn serialize_struct(
+        self,
+        _name: &'static str,
+        len: usize,
+    ) -> Result<Self::SerializeStruct, Self::Error> {
+        self.serialize_map(Some(len))
+    }
+
+    fn serialize_struct_variant(
+        self,
+        _name: &'static str,
+        _variant_index: u32,
+        variant: &'static str,
+        _len: usize,
+    ) -> Result<Self::SerializeStructVariant, Self::Error> {
+        let mut map = Map::new();
+        map.insert(variant.to_string(), Value::Table(Map::new()));
+        Ok(SerializeMap {
+            map: Map::new(),
+            next_key: None,
+        })
+    }
+
+    fn serialize_i128(self, v: i128) -> Result<Self::Ok, Self::Error> {
+        let _ = v;
+        Err(serde::ser::Error::custom("i128 is not supported"))
+    }
+
+    fn serialize_u128(self, v: u128) -> Result<Self::Ok, Self::Error> {
+        let _ = v;
+        Err(serde::ser::Error::custom("u128 is not supported"))
+    }
+
+    // fn collect_seq<I>(self, iter: I) -> Result<Self::Ok, Self::Error>
+    // where
+    //     I: IntoIterator,
+    //     <I as IntoIterator>::Item: Serialize,
+    // {
+    //     let mut iter = std::iter.into_iter();
+    //     let mut serializer =
+    // tri!(self.serialize_seq(iterator_len_hint(&iter)));     tri!(iter.
+    // try_for_each(|item| serializer.serialize_element(&item)));
+    //     serializer.end()
+    // }
+
+    // fn collect_map<K, V, I>(self, iter: I) -> Result<Self::Ok, Self::Error>
+    // where
+    //     K: Serialize,
+    //     V: Serialize,
+    //     I: IntoIterator<Item = (K, V)>,
+    // {
+    //     let mut iter = std::iter.into_iter();
+    //     let mut serializer =
+    // tri!(self.serialize_map(iterator_len_hint(&iter)));     tri!(iter.
+    // try_for_each(|(key, value)| serializer.serialize_entry(&key, &value)));
+    //     serializer.end()
+    // }
+
+    fn collect_str<T>(self, value: &T) -> Result<Self::Ok, Self::Error>
+    where
+        T: ?Sized + std::fmt::Display,
+    {
+        self.serialize_str(&value.to_string())
+    }
+
+    fn is_human_readable(&self) -> bool {
+        true
+    }
+}
+
+pub struct SerializeSeq {
+    vec: Vec<Value>,
+}
+
+impl serde::ser::SerializeSeq for SerializeSeq {
+    type Ok = Value;
+    type Error = RealmError;
+
+    fn serialize_element<T>(&mut self, value: &T) -> Result<(), Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        self.vec.push(value.serialize(&mut ValueSerializer)?);
+        Ok(())
+    }
+
+    fn end(self) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Array(self.vec))
+    }
+}
+
+impl serde::ser::SerializeTuple for SerializeSeq {
+    type Ok = Value;
+    type Error = RealmError;
+
+    fn serialize_element<T>(&mut self, value: &T) -> Result<(), Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        self.vec.push(value.serialize(&mut ValueSerializer)?);
+        Ok(())
+    }
+
+    fn end(self) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Array(self.vec))
+    }
+}
+
+impl serde::ser::SerializeTupleStruct for SerializeSeq {
+    type Ok = Value;
+    type Error = RealmError;
+
+    fn serialize_field<T>(&mut self, value: &T) -> Result<(), Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        self.vec.push(value.serialize(&mut ValueSerializer)?);
+        Ok(())
+    }
+
+    fn end(self) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Array(self.vec))
+    }
+}
+
+impl serde::ser::SerializeTupleVariant for SerializeSeq {
+    type Ok = Value;
+    type Error = RealmError;
+
+    fn serialize_field<T>(&mut self, value: &T) -> Result<(), Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        self.vec.push(value.serialize(&mut ValueSerializer)?);
+        Ok(())
+    }
+
+    fn end(self) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Array(self.vec))
+    }
+}
+
+pub struct SerializeMap {
+    map: Map<String, Value>,
+    next_key: Option<String>,
+}
+
+impl serde::ser::SerializeMap for SerializeMap {
+    type Ok = Value;
+    type Error = RealmError;
+
+    fn serialize_key<T>(&mut self, key: &T) -> Result<(), Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        if let Value::String(key) = key.serialize(&mut ValueSerializer)? {
+            self.next_key = Some(key);
+            Ok(())
+        } else {
+            Err(serde::ser::Error::custom("Map key must be a string"))
+        }
+    }
+
+    fn serialize_value<T>(&mut self, value: &T) -> Result<(), Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        if let Some(key) = self.next_key.take() {
+            self.map.insert(key, value.serialize(&mut ValueSerializer)?);
+            Ok(())
+        } else {
+            Err(serde::ser::Error::custom("Value serialized before key"))
+        }
+    }
+
+    fn end(self) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Table(self.map))
+    }
+}
+
+impl serde::ser::SerializeStruct for SerializeMap {
+    type Ok = Value;
+    type Error = RealmError;
+
+    fn serialize_field<T>(
+        &mut self,
+        key: &'static str,
+        value: &T,
+    ) -> Result<(), Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        self.map
+            .insert(key.to_string(), value.serialize(&mut ValueSerializer)?);
+        Ok(())
+    }
+
+    fn end(self) -> Result<Self::Ok, Self::Error> {
+        Ok(Value::Table(self.map))
+    }
+}
+
+impl serde::ser::SerializeStructVariant for SerializeMap {
+    type Ok = Value;
+    type Error = RealmError;
+
+    fn serialize_field<T>(
+        &mut self,
+        key: &'static str,
+        value: &T,
+    ) -> Result<(), Self::Error>
+    where
+        T: ?Sized + Serialize,
+    {
+        self.map
+            .insert(key.to_string(), value.serialize(&mut ValueSerializer)?);
+        Ok(())
+    }
+
+    fn end(self) -> Result<Self::Ok, Self::Error> {
+        todo!()
+    }
+}
+#[cfg(test)]
+mod test {
+    use serde::{Deserialize, Serialize};
+
+    use super::*;
+
+    #[test]
+    fn test_struct() {
+        #[derive(Debug, Serialize, Deserialize, PartialEq, Eq)]
+        struct Test {
+            int: u32,
+            seq: Vec<String>,
+        }
+
+        let test = Test {
+            int: 1,
+            seq: vec!["a".to_string(), "b".to_string()],
+        };
+        let config = Value::try_serialize(&test).unwrap();
+
+        let actual: Test = config.try_deserialize().unwrap();
+        assert_eq!(test, actual);
+    }
+
+    #[test]
+    fn test_nest() {
+        let val = serde_json::json! { {
+            "top": {
+                "num": 1,
+                "array": [2],
+                "nested": [[3,4]],
+                "deep": [{
+                    "yes": true,
+                }],
+                "mixed": [
+                    { "boolish": false, },
+                    42,
+                    ["hi"],
+                    { "inner": 66 },
+                    23,
+                ],
+            }
+        } };
+        let config = Value::try_serialize(&val).unwrap();
+        let output: serde_json::Value = config.try_deserialize().unwrap();
+        assert_eq!(val, output);
+    }
+}